refactor: 카드 셀 클래스의 init이 잘 보이기 위해 각각의 하위뷰들의 클래스를 만들고 적용함
